Bring timestamp back to usize -- curse you, large_integer_atomics!

+++ src/timestamp.rs
@@ -1,13 +1,13 @@
 use std::sync::atomic::{AtomicUsize, Ordering};
 
-pub type Timestamp = u64;
+pub type Timestamp = usize;
 
 lazy_static! {
     static ref LAST_TIMESTAMP: AtomicUsize = AtomicUsize::new(0);
 }
 
 pub fn get_timestamp() -> Timestamp {
-    LAST_TIMESTAMP.fetch_add(1, Ordering::SeqCst) as u64
+    LAST_TIMESTAMP.fetch_add(1, Ordering::SeqCst)
 }
 
 #[test]
